+++ src/main/scala/gcli/GoogleQueryCmd.scala
@@ -7,13 +7,14 @@ package gcli
 class GoogleQueryCmd {
   var startCommand: String = _
   var tld: String = _
+  var googleSearchType: String = _
   var limitToSite: String = _
   var q: String = _
   var timeFilter: String = _
   var noop: Boolean = false
 
   /** Returns the shell command for the given values of this object. */
-  def buildCommand: String = s"$startCommand https://www.google.$tld/?#q=$limitToSite$q$timeFilter"
+  def buildCommand: String = s"$startCommand https://www.google.$tld/?#q=$limitToSite$q$timeFilter$googleSearchType"
 
   def isValid: Boolean = {
     // TODO implement